requirements.txt
tinkoff-investments
python-dotenv
ta